requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==14.1
google-generativeai==0.8.3
graphviz==0.20.3
Pillow==11.0.0
python-dotenv==1.0.1
aiofiles==24.1.0

